fix!: Update testing helpers
The testing package used to contain some helper methods and types that
could be used inside of policy unit tests.

The module defined objects that are now defined inside of the `protocol`
package thanks to us leveraging easyjson.

The `testing` code has been updated to:
  * Remove types that are now declared inside of the `protocol` package
  * Update the functions to make use of the new types used by policies

This is a breaking API change that is going to break old policies unit
tests.

Signed-off-by: Flavio Castelli <fcastelli@suse.com>