chore(deps): use proper version of strfmt library
For some reason the `go.mod` file was requiring a version of the strfmt
that didn't have a semantic version.

This lead to parsing errors inside of renovate bot, causing all our Go
policies to be not be handled by the automatic dependency updater.

This is the error that renovate bot was facing:

```
go: github.com/kubewarden/policy-sdk-go@v0.2.4 requires
	github.com/go-openapi/strfmt@v0.0.0-00010101000000-000000000000: invalid version: unknown revision 000000000000
```

This is now fixed by this commit.

Signed-off-by: Flavio Castelli <fcastelli@suse.com>