fix: support changes done by TinyGo 0.32.0
Starting from TinyGo 0.32.0 release, the `wasi` target has been renamed
to `wasip1` to match the one used by the Go official compiler (aka
`gc`).

This change broke some of our build tags. For example, the `protocol`
file was no longer built by our wapc policies, causing errors when the
policy was being annotated.

This commit changes the build tags to use the `tinygo` build tag to
identify builds done by TinyGo. The assumption is: we use TinyGo only to
build WebAssembly modules (not to build x86_64 ones), hence we can
safely assume that, when this build tag is found, we're building a
policy that is meant to be used with WaPC.

Policies being built with older versions of TinyGo will keep working
fine, since this build tag has seems to have always been part of TinyGo

Signed-off-by: Flavio Castelli <fcastelli@suse.com>